moving view models into views/
